resume_screening_tool/requirements.txt
streamlit
scikit-learn
nltk
PyPDF2
python-docx
pandas
